notify finction TS cont.

+++ functions/index.ts
@@ -27,7 +27,7 @@ export const notify = onRequest(
     })
 
     if (registrationTokens.length === 0) {
-      res.status(200).send('No subscribers found')
+      res.status(200).send('No subscribers found error')
       return
     }
 
@@ -44,14 +44,27 @@ export const notify = onRequest(
       getMessaging()
         .sendEachForMulticast(message)
         .then((response) => {
-          if (response.failureCount > 0) {
-            const failedTokens: string[] = []
-            response.responses.forEach((resp, idx) => {
-              if (resp && !resp.success && idx < registrationTokens.length) {
+          const failedTokens: string[] = []
+          const successTokens: string[] = []
+          response.responses.forEach((resp, idx) => {
+            if (resp && idx < registrationTokens.length) {
+              if (!resp.success) {
                 failedTokens.push(registrationTokens[idx]!)
+              } else {
+                successTokens.push(registrationTokens[idx]!)
               }
-            })
+            }
+          })
+          if (failedTokens.length > 0) {
             failedTokens.forEach((token) => removeToken(token))
+            // logger.info(`Removed ${failedTokens.length} failed tokens`)
+          }
+          if (successTokens.length > 0) {
+            successTokens.forEach((token) => messageSent(token))
+            // logger.info(`Successfully sent to ${successTokens.length} tokens`)
+          } else if (successTokens.length === 0) {
+            res.status(200).send('No active subscribers found')
+            logger.info(`No active subscribers found. No message sent`)
           }
         })
     } catch (error) {
@@ -66,5 +79,13 @@ const removeToken = async (token: string): Promise<void> => {
   const docRef = getFirestore().collection('Device').doc(token)
   const doc = await docRef.get()
   logger.info(`Remove token for ${doc.data()?.email}`)
+  await docRef.delete()
+}
+
+const messageSent = async (token: string): Promise<void> => {
+  if (token === undefined) return
+  const docRef = getFirestore().collection('Device').doc(token)
+  const doc = await docRef.get()
+  logger.info(`Message sent to ${doc.data()?.email}`)
   await docRef.delete()
 }
